requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
simplejwt==2.0.1
sqlparse==0.5.5
typing==3.7.4.3
tzdata==2025.3
